Remove "wait for heat up" G-code
Since the heating code is already inserted by Cura before any of this
code runs (and it has to be, until Ultimaker/Cura#1296 is resolved), the
only function this has is to move the print head - at an unsettlingly
fast speed - to the "prime position" (which, as we haven't set it here,
is set to 0,0, ie. the center of the bed) right before starting the
print, which isn't useful. In lieu of a proper in-code heating procedure
that could make use of this priming, better to just start from the back
corner where the extruders are zeroed to.